changed all double to LDBLE
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/branches/rename_cpp@6038 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ PHRQ_io_output.cpp
@@ -116,7 +116,7 @@ fpunchf_user(int user_index, const char *format, double d)
 				(user_index - user_punch_count_headings) + 1);
 		name = buffer;
 	}
-	if (phrq_io) phrq_io->fpunchf(name, format, d);
+	if (phrq_io) phrq_io->fpunchf(name, format, (double) d);
 }
 
 void Phreeqc::
